+++ src/modules/server/serverMessages.js
@@ -4,9 +4,16 @@
 
 // messages for redirect incoming events from server to IPC-Main
 module.exports = {
+  // ********************************
+  // ipc-server -> ipc-main
+  // ********************************
   UPDATE_CONNECTION_STATUS: "update-connection-status",
-  UPDATE_SETS: "update-sets",
-  // TODO: improve naming
   STATE_REQUEST: "state-request",
-  STATE_RESPONSE: "state-response"
+  UPDATE_SETS_REQUEST: "update-sets-request",
+
+  // ********************************
+  // ipc-main -> ipc-server
+  // ********************************
+  STATE_RESPONSE: "state-response",
+  UPDATE_SETS_RESPONSE: "update-sets-response"
 };
